backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
groq
python-dotenv
email-validator